app/backend/requirements.txt
Flask==2.0.2
gunicorn
requests
openai
line-bot-sdk
requests
python-dotenv
azure-identity
azure-search-documents
azure-cosmos
